feat: Phase 13.0 — Meta Reasoning & Self Reflection Engine
Implement the final intelligence layer before the Verification Fabric
executes a plan. The Meta Reasoner reflects, critiques, validates,
simulates, generates alternatives, verifies evidence, and refines plans.

## Package: backend/meta_reasoning/ (18 modules)

| Module | Purpose |
|--------|---------|
| schemas.py | ReasoningResult, SimulationResult, QualityReport + 3 enums |
| exceptions.py | MetaReasoningError hierarchy (4 types) |
| events.py | 12 event types for reasoning lifecycle |
| reasoning_trace.py | Persistent reasoning audit trail |
| plan_validator.py | 6-check validation (deps, caps, perms, budget, risk, policies) |
| reflection_engine.py | 6-dimension self-reflection (goal, reasoning, planning quality) |
| critique_engine.py | 7-detector critique (logic gaps, assumptions, contradictions) |
| alternative_generator.py | 5-strategy alternatives + ranking |
| hypothesis_engine.py | Hypothesis generation, evaluation, ranking |
| counterfactual_engine.py | 6 what-if scenario analyzers |
| simulation_engine.py | Dry-run simulation (latency, cost, tokens, failure, resources) |
| evidence_engine.py | 4-domain evidence verification + confidence scoring |
| plan_refiner.py | 6 refinement strategies + improvement diffing |
| reasoning_memory.py | Experience store with success/failure analytics |
| uncertainty_engine.py | 4 uncertainty detectors + overall score |
| quality_estimator.py | 8-dimension quality scoring |
| meta_reasoner.py | 10-stage async orchestrator with deadlock protection |
| __init__.py | Package exports |

## Meta Reasoning Pipeline

Executive Brain → Reflection → Critique → Evidence Verification
→ Counterfactual Analysis → Simulation → Quality Estimation
→ Confidence/Uncertainty → Plan Refinement → Verdict → Verification Fabric

## Deadlock Protection

- Maximum 3 refinement iterations
- After 3 iterations: BLOCKED + persist all analysis + human intervention
- Meta Reasoner NEVER continues automatically after BLOCKED

## API Endpoints (api/meta_reasoning.py)

| Method | Endpoint | Description |
|--------|----------|-------------|
| POST | /api/v1/meta-reasoning/reason | Full reasoning pipeline |
| GET | /api/v1/meta-reasoning/status | Reasoner status |
| GET | /api/v1/meta-reasoning/trace/{id} | Reasoning trace |
| GET | /api/v1/meta-reasoning/events | Recent events |
| GET | /api/v1/meta-reasoning/memory/stats | Memory statistics |

## Tests: 110 new (530 total)

## Validation
- Ruff: 0 errors (263 files)
- Mypy: 0 errors (242 source files)
- Bandit: 0 findings
- Pytest: 530/530 passed (3.97s)

+++ backend/meta_reasoning/simulation_engine.py
@@ -60,15 +60,15 @@ class SimulationEngine:
         mode = plan.get("execution_mode", "sequential")
 
         if mode == "parallel":
-            return base_latency + (task_count * 100.0)
-        return base_latency + (task_count * 200.0)
+            return float(base_latency + (task_count * 100.0))
+        return float(base_latency + (task_count * 200.0))
 
     def _estimate_cost(self, plan: dict) -> float:
         """Estimate total execution cost."""
         cost_info = plan.get("cost", {})
         base_cost = cost_info.get("total_estimated", 0.01)
         tasks = plan.get("tasks", [])
-        return base_cost + (len(tasks) * 0.005)
+        return float(base_cost + (len(tasks) * 0.005))
 
     def _estimate_tokens(self, plan: dict) -> int:
         """Estimate total token usage."""
@@ -85,7 +85,7 @@ class SimulationEngine:
         task_factor = min(len(tasks) * 0.02, 0.3)
         mode = plan.get("execution_mode", "sequential")
         mode_factor = 0.05 if mode == "parallel" else 0.0
-        return min(base_risk + task_factor + mode_factor, 1.0)
+        return float(min(base_risk + task_factor + mode_factor, 1.0))
 
     def _estimate_resource_usage(self, plan: dict) -> dict[str, Any]:
         """Estimate resource usage during execution."""
@@ -126,13 +126,15 @@ class SimulationEngine:
         for i, task in enumerate(tasks):
             step_latency = 200.0
             cumulative_ms += step_latency
-            trace_steps.append({
-                "step": i + 1,
-                "task": task,
-                "status": "simulated_success",
-                "latency_ms": step_latency,
-                "cumulative_ms": cumulative_ms,
-            })
+            trace_steps.append(
+                {
+                    "step": i + 1,
+                    "task": task,
+                    "status": "simulated_success",
+                    "latency_ms": step_latency,
+                    "cumulative_ms": cumulative_ms,
+                }
+            )
 
         return {
             "plan_id": plan.get("plan_id", "unknown"),
